requirements.txt
torch>=2.0.0
numpy>=1.24.0
pyyaml>=6.0
datasets>=2.20.0
